+++ tornado-drivers/opencl/src/main/java/uk/ac/manchester/tornado/drivers/opencl/power/OCLNvidiaPowerMetricHandler.java
@@ -28,12 +28,13 @@ import uk.ac.manchester.tornado.drivers.opencl.OCLDeviceContext;
 import uk.ac.manchester.tornado.drivers.opencl.exceptions.OCLException;
 import uk.ac.manchester.tornado.runtime.common.TornadoLogger;
 
-public class OCLNvidiaPowerMetric implements PowerMetric {
+public class OCLNvidiaPowerMetricHandler implements PowerMetric {
 
     private final OCLDeviceContext deviceContext;
     private final TornadoLogger logger;
+    private long[] oclDevice = new long[1];
 
-    public OCLNvidiaPowerMetric(OCLDeviceContext deviceContext) {
+    public OCLNvidiaPowerMetricHandler(OCLDeviceContext deviceContext) {
         this.deviceContext = deviceContext;
         this.logger = new TornadoLogger(this.getClass());
         initializePowerLibrary();
@@ -49,24 +50,16 @@ public class OCLNvidiaPowerMetric implements PowerMetric {
     public void initializePowerLibrary() {
         try {
             clNvmlInit();
+            clNvmlDeviceGetHandleByIndex(this.deviceContext.getDevice().getIndex(), this.oclDevice);
         } catch (OCLException e) {
             logger.error(e.getMessage());
         }
     }
 
     @Override
-    public void getHandleByIndex(long[] device) {
+    public void getPowerUsage(long[] powerUsage) {
         try {
-            clNvmlDeviceGetHandleByIndex(this.deviceContext.getDevice().getIndex(), device);
-        } catch (OCLException e) {
-            logger.error(e.getMessage());
-        }
-    }
-
-    @Override
-    public void getPowerUsage(long[] device, long[] powerUsage) {
-        try {
-            clNvmlDeviceGetPowerUsage(device, powerUsage);
+            clNvmlDeviceGetPowerUsage(this.oclDevice, powerUsage);
         } catch (OCLException e) {
             logger.error(e.getMessage());
         }