+++ tornado-drivers/ptx/src/main/java/uk/ac/manchester/tornado/drivers/ptx/graal/lir/PTXArithmeticTool.java
@@ -31,13 +31,16 @@ import jdk.graal.compiler.core.common.LIRKind;
 import jdk.graal.compiler.core.common.calc.FloatConvert;
 import jdk.graal.compiler.core.common.memory.MemoryExtendKind;
 import jdk.graal.compiler.core.common.memory.MemoryOrderMode;
+import jdk.graal.compiler.lir.ConstantValue;
 import jdk.graal.compiler.lir.LIRFrameState;
 import jdk.graal.compiler.lir.Variable;
 import jdk.graal.compiler.lir.gen.ArithmeticLIRGenerator;
+import jdk.vm.ci.meta.JavaConstant;
 
 import jdk.vm.ci.meta.PlatformKind;
 import jdk.vm.ci.meta.Value;
 import jdk.vm.ci.meta.ValueKind;
+import uk.ac.manchester.tornado.drivers.common.code.CodeUtil;
 import uk.ac.manchester.tornado.drivers.common.logging.Logger;
 import uk.ac.manchester.tornado.drivers.ptx.graal.PTXLIRKindTool;
 import uk.ac.manchester.tornado.drivers.ptx.graal.asm.PTXAssembler;
@@ -197,12 +200,6 @@ public class PTXArithmeticTool extends ArithmeticLIRGenerator {
     @Override
     public Value emitSignExtend(Value inputVal, int fromBits, int toBits) {
         Logger.traceBuildLIR(Logger.BACKEND.PTX, "emitSignExtend inputVal=%s fromBits=%d toBits=%d", inputVal, fromBits, toBits);
-        return emitZeroExtend(inputVal, fromBits, toBits);
-    }
-
-    @Override
-    public Value emitZeroExtend(Value inputVal, int fromBits, int toBits) {
-        Logger.traceBuildLIR(Logger.BACKEND.PTX, "emitZeroExtend inputVal=%s fromBits=%d toBits=%d", inputVal, fromBits, toBits);
         PTXLIRKindTool kindTool = getGen().getLIRKindTool();
         PTXKind kind = (PTXKind) inputVal.getPlatformKind();
         LIRKind toKind;
@@ -217,6 +214,30 @@ public class PTXArithmeticTool extends ArithmeticLIRGenerator {
         Variable result = getGen().newVariable(toKind);
 
         getGen().emitMove(result, inputVal);
+        return result;
+
+    }
+
+    @Override
+    public Value emitZeroExtend(Value inputVal, int fromBits, int toBits) {
+        Logger.traceBuildLIR(Logger.BACKEND.PTX, "emitZeroExtend inputVal=%s fromBits=%d toBits=%d", inputVal, fromBits, toBits);
+        PTXLIRKindTool kindTool = getGen().getLIRKindTool();
+        PTXKind kind = (PTXKind) inputVal.getPlatformKind();
+        LIRKind toKind;
+        if (kind.isInteger()) {
+            toKind = kindTool.getUnsignedIntegerKind(toBits);
+        } else if (kind.isFloating()) {
+            toKind = kindTool.getFloatingKind(toBits);
+        } else {
+            throw shouldNotReachHere();
+        }
+
+        Variable signExtendedValue = getGen().newVariable(toKind);
+        getGen().emitMove(signExtendedValue, inputVal);
+
+        // Apply a bitwise mask in order to avoid sign extension and instead zero extend the value.
+        ConstantValue mask = new ConstantValue(toKind, JavaConstant.forIntegerKind(CodeUtil.javaKindFromBitSize(toBits, kind.isFloating()), (1L << fromBits) - 1));
+        Variable result = emitBinaryAssign(PTXBinaryOp.BITWISE_AND, toKind, signExtendedValue, mask);
         return result;
     }