+++ tornado-unittests/src/main/java/uk/ac/manchester/tornado/unittests/kernelcontext/api/KernelContextWorkGroupTests.java
@@ -40,7 +40,7 @@ import uk.ac.manchester.tornado.unittests.common.TornadoTestBase;
  * How to run.
  * </p>
  * <code>
- * tornado-test --threadInfo --printKernel --fast -V uk.ac.manchester.tornado.unittests.kernelcontext.api.KernelContextWorkGroupTests
+ * tornado-test --threadInfo --printKernel -V uk.ac.manchester.tornado.unittests.kernelcontext.api.KernelContextWorkGroupTests
  * </code>
  */
 public class KernelContextWorkGroupTests extends TornadoTestBase {