+++ docs/source/CHANGELOG.rst
@@ -5,12 +5,41 @@ TornadoVM Changelog
 
 This file summarizes the new features and major changes for each *TornadoVM* version.
 
+TornadoVM 1.0.8
+---------------
+30th September 2024
+
+Improvements
+~~~~~~~~~~~~
+
+- `#565 <https://github.com/beehive-lab/TornadoVM/pull/565>`_: New API call in the Execution Plan to log/trace the executed configuration plans.
+- `#563 <https://github.com/beehive-lab/TornadoVM/pull/563>`_: Expand the TornadoVM profiler with Level Zero Sysman Energy Metrics.
+- `#559 <https://github.com/beehive-lab/TornadoVM/pull/559>`_: Refactoring Power Metric handlers for PTX and OpenCL.
+- `#548 <https://github.com/beehive-lab/TornadoVM/pull/548>`_: Benchmarking improvements.
+- `#549 <https://github.com/beehive-lab/TornadoVM/pull/549>`_: Prebuilt API tests added using multiple backend-setup.
+- Add internal tests for monitoring memory management `(link) <https://github.com/beehive-lab/TornadoVM/commit/0644225a641bd859372743b59d46c6c9a4613337>`_.
+
+Compatibility
+~~~~~~~~~~~~~
+- `#561 <https://github.com/beehive-lab/TornadoVM/pull/561>`_: Build for OSx 14.6 and OSx 15 fixed.
+
+Bug Fixes
+~~~~~~~~~
+
+- `#564 <https://github.com/beehive-lab/TornadoVM/pull/564>`_: Jenkins configuration fixed to run KFusion per backend.
+- `#562 <https://github.com/beehive-lab/TornadoVM/pull/562>`_: Warmup action from the Execution Plan fixed to run with correct internal IDs.
+- `#557 <https://github.com/beehive-lab/TornadoVM/pull/557>`_: Shared Execution Plans Context fixed.
+- `#553 <https://github.com/beehive-lab/TornadoVM/pull/553>`_: OpenCL compiler flags for Intel Integrated GPUs fixed.
+- `#552 <https://github.com/beehive-lab/TornadoVM/pull/552>`_: Fixed runtime to select any device among multiple SPIR-V devices.
+- Fixed zero extend arithmetic operations: `link <https://github.com/beehive-lab/TornadoVM/commit/ea7b60263072ba0299da205cb920d0c68b3d1749>`_
+
+
 TornadoVM 1.0.7
 ----------------
 30th August 2024
 
 Improvements
-~~~~~~~~~~~~~~~~~~
+~~~~~~~~~~~~
 
 - `#468 <https://github.com/beehive-lab/TornadoVM/pull/468>`_: Cleanup Abstract Metadata Class.
 - `#473 <https://github.com/beehive-lab/TornadoVM/pull/473>`_: Add maven plugin to build TornadoVM source for the releases.
@@ -27,7 +56,7 @@ Improvements
 - `#542 <https://github.com/beehive-lab/TornadoVM/pull/542>`_: Tagged LevelZero JNI and Beehive Toolkit dependencies added in the build and installer.
 
 Compatibility
-~~~~~~~~~~~~~~~~~~
+~~~~~~~~~~~~~
 
 - `#465 <https://github.com/beehive-lab/TornadoVM/pull/465>`_: Support for JDK 22 and GraalVM 24.0.2.
 - `#486 <https://github.com/beehive-lab/TornadoVM/pull/486>`_: Temurin for Windows added in the list of supported JDKs.